fix equlas
