+++ hutool-cron/src/main/java/cn/hutool/cron/pattern/parser/DayOfMonthValueParser.java
@@ -2,8 +2,6 @@ package cn.hutool.cron.pattern.parser;
 
 import cn.hutool.cron.CronException;
 
-import java.util.Objects;
-
 /**
  * 每月的几号值处理<br>
  * 每月最多31天，32和“L”都表示最后一天
@@ -19,7 +17,7 @@ public class DayOfMonthValueParser extends SimpleValueParser {
 
 	@Override
 	public int parse(String value) throws CronException {
-		if ("L".equalsIgnoreCase(value) || Objects.equals("32",value)) {// 每月最后一天
+		if ("L".equalsIgnoreCase(value) || "32".equals(value)) {// 每月最后一天
 			return 32;
 		} else {
 			return super.parse(value);
