+++ hutool-db/src/main/java/cn/hutool/db/sql/SqlFormatter.java
@@ -126,7 +126,7 @@ public class SqlFormatter {
 					values();
 				} else if ("on".equals(this.lcToken)) {
 					on();
-				} else if ((this.afterBetween) && (Objects.equals(this.lcToken,"and"))) {
+				} else if ((this.afterBetween) && ("and".equals(this.lcToken))) {
 					misc();
 					this.afterBetween = false;
 				} else if (LOGICAL.contains(this.lcToken)) {
@@ -213,8 +213,8 @@ public class SqlFormatter {
 			out();
 			this.indent += 1;
 			newline();
-			this.parenCounts.addLast(new Integer(this.parensSinceSelect));
-			this.afterByOrFromOrSelects.addLast(Boolean.valueOf(this.afterByOrSetOrFromOrSelect));
+			this.parenCounts.addLast(this.parensSinceSelect);
+			this.afterByOrFromOrSelects.addLast(this.afterByOrSetOrFromOrSelect);
 			this.parensSinceSelect = 0;
 			this.afterByOrSetOrFromOrSelect = true;
 		}
@@ -268,8 +268,8 @@ public class SqlFormatter {
 			this.parensSinceSelect -= 1;
 			if (this.parensSinceSelect < 0) {
 				this.indent -= 1;
-				this.parensSinceSelect = ((Integer) this.parenCounts.removeLast()).intValue();
-				this.afterByOrSetOrFromOrSelect = ((Boolean) this.afterByOrFromOrSelects.removeLast()).booleanValue();
+				this.parensSinceSelect = this.parenCounts.removeLast();
+				this.afterByOrSetOrFromOrSelect = this.afterByOrFromOrSelects.removeLast();
 			}
 			if (this.inFunction > 0) {
 				this.inFunction -= 1;
