+++ hutool-extra/src/main/java/cn/hutool/extra/ssh/Sftp.java
@@ -15,7 +15,6 @@ import java.io.File;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Vector;
 
 /**
@@ -312,7 +311,7 @@ public class Sftp extends AbstractFtp {
 		String fileName;
 		for (LsEntry entry : list) {
 			fileName = entry.getFilename();
-			if (false == Objects.equals(fileName,".") && false == Objects.equals(fileName,"..")) {
+			if (false == ".".equals(fileName) && false == "..".equals(fileName)) {
 				if (entry.getAttrs().isDir()) {
 					delDir(fileName);
 				} else {
