+++ hutool-http/src/main/java/cn/hutool/http/HttpRequest.java
@@ -381,7 +381,7 @@ public class HttpRequest extends HttpBase<HttpRequest> {
 			return !httpVersion.equalsIgnoreCase(HTTP_1_0);
 		}
 
-		return !"close".equalsIgnoreCase(connection);
+		return false == "close".equalsIgnoreCase(connection);
 	}
 
 	/**
